requirements.txt
opencv-python
pillow
pytesseract
numpy
pandas
Flask
pytesseract